Movie view layout is squished in portrait mode – workaround: hold your iPad landscape, since that makes sense for watching a movie anyway… WORKAROUND – view changed to only show landscape left or right
Home screen top-right menu button is unimplemented, needs to link to search options. WORKAROUND – button links to menu of options, including credits and this wiki

There needs to be an Actor page that shows the actor and bio etc, as well as the list of movies they are in. WORKAROUND – button links to Netflix site bio for the actor